+++ pennylane_lightning/lightning_qubit/lightning_qubit2.py
@@ -246,7 +246,8 @@ class LightningQubit2(Device):
                     f"The {kernel_name} is not supported and currently "
                     "only 'Local' and 'NonZeroRandom' kernels are supported."
                 )
-            if num_burnin >= shots:
+            shots = shots if isinstance(shots, Sequence) else [shots]
+            if any(num_burnin >= s for s in shots):
                 raise ValueError("Shots should be greater than num_burnin.")
             self._kernel_name = kernel_name
             self._num_burnin = num_burnin